code_rabbit/coderabbit_dashboard_python/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.9.2
aiosqlite==0.20.0
PyGithub==2.4.0
aiosqlite==0.20.0
